fix: eagerly resolve whether to show missing leave report notice
Lazy resolution amounted to a bug such that the div would start showing as a side effect of rendering the leave reports table (so, navigating to that tab).
This eliminated much of the value of a notice in helping employees to know they need to be navigating to that tab.