HRSPLT-431 open HRS Fluid Time in new tab or window
In case where launching MyUW Time and Absence instead launches HRS Self-service "Fluid Time", make that happen in a new "window" (often, tab, depending on browser), and return the MyUW window to the MyUW home page. This is to keep with MyUW's local style that launches of applications external MyUW that don't feel like MyUW apps, happens in a new tab.